% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureBuildCho.R
\name{signatureBuildCho}
\alias{signatureBuildCho}
\title{Build the standard input file for the stress signatures from Bryan Chambers}
\usage{
signatureBuildCho()
}
\value{
No output.
}
\description{
Build the standard input file for the stress signatures from Bryan Chambers
}
