% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureBuildCorton.R
\name{signatureBuildCorton}
\alias{signatureBuildCorton}
\title{Build the standard input file for the Corton signatures}
\usage{
signatureBuildCorton()
}
\value{
No output.
}
\description{
Build the standard input file for the Corton signatures
}
