% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureBuildDorothea.R
\name{signatureBuildDorothea}
\alias{signatureBuildDorothea}
\title{Build the standard input file for the Dorothea signatures}
\usage{
signatureBuildDorothea()
}
\value{
No output.
}
\description{
Build the standard input file for the Dorothea signatures
}
