% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureBuilder v2.R
\name{signatureBuilder.v2}
\alias{signatureBuilder.v2}
\title{Create the files needed for the signature calculations before adding random genes}
\usage{
signatureBuilder.v2(min.ngene = 10, max.ngene = 1e+05)
}
\arguments{
\item{min.ngene}{Signatures will only be saved if the number of genes is >= this value}

\item{max.ngene}{Signatures will only be saved if the number of genes is <= this value}
}
\value{
No output.
}
\description{
Create the files needed for the signature calculations before adding random genes
}
