% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureBuildStress.R
\name{signatureBuildStress}
\alias{signatureBuildStress}
\title{Build the standard input file for the stress signatures from Bryan Chambers}
\usage{
signatureBuildStress()
}
\value{
No output.
}
\description{
Build the standard input file for the stress signatures from Bryan Chambers
}
