% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureDBupdate.R
\name{signatureDBupdate}
\alias{signatureDBupdate}
\title{Save the current set of signatures before rebuilding}
\usage{
signatureDBupdate(do.save = T)
}
\value{
No output.
}
\description{
Save the current set of signatures before rebuilding
}
